Fix minimum MSM filter, generalise MzFilter and use for both
